% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BigColSums}
\alias{BigColSums}
\title{Dispatch function for BigColSums}
\usage{
BigColSums(pBigMat)
}
\arguments{
\item{pBigMat}{SEXP container for the external pointer to the BigMatrix
object}
}
\description{
Dispatch function for BigColSums
}

